modify function locations
